requirements.txt
Pillow
playwright
beautifulsoup4
# spidev
numpy
gpiozero